+++ src/expr/src/vector_op/mod.rs
@@ -25,6 +25,7 @@ pub mod date_trunc;
 pub mod exp;
 pub mod extract;
 pub mod format_type;
+pub mod trigonometric;
 pub mod jsonb_info;
 pub mod length;
 pub mod like;
